# openapi.yaml
openapi: 3.0.3
info:
  title: Your API
  version: 1.0.0
  description: REST API for authentication and account management (and more).
  contact:
    name: Team
    email: support@example.com
  license:
    name: Proprietary
paths:
...
  /api/v1/investors/{id}/:
    get:
      operationId: investors_retrieve
      description: |-
        ViewSet for managing Investor instances.
        Optimized with select_related to avoid N+1 queries when fetching related user, industry, and location.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Investor.
        required: true
      tags:
      - investors
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investor'
          description: ''
    put:
      operationId: investors_update
      description: |-
        ViewSet for managing Investor instances.
        Optimized with select_related to avoid N+1 queries when fetching related user, industry, and location.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Investor.
        required: true
      tags:
      - investors
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestorRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InvestorRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InvestorRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investor'
          description: ''
    patch:
      operationId: investors_partial_update
      description: |-
        ViewSet for managing Investor instances.
        Optimized with select_related to avoid N+1 queries when fetching related user, industry, and location.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Investor.
        required: true
      tags:
      - investors
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedInvestorRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedInvestorRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedInvestorRequest'
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investor'
          description: ''
    delete:
      operationId: investors_destroy
      description: |-
        ViewSet for managing Investor instances.
        Optimized with select_related to avoid N+1 queries when fetching related user, industry, and location.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Investor.
        required: true
      tags:
      - investors
      security:
      - bearerAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    Investor:
      type: object
      description: |-
        Serializer for the Investor model.
        Includes all fields defined in the abstract Company base class and Investor-specific fields.
      properties:
        id:
          type: integer
          readOnly: true
        user:
          type: integer
          readOnly: true
          description: The user who owns this investor
        industry:
          type: integer
        company_name:
          type: string
          maxLength: 254
        location:
          type: integer
        logo:
          type: string
          format: uri
          nullable: true
        description:
          type: string
        website:
          type: string
          format: uri
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        founded_year:
          type: integer
          minimum: 1900
          maximum: 2025
        team_size:
          type: integer
          minimum: 1
        stage:
          enum:
          - idea
          - mvp
          - launch
          - scale
          - exit
          type: string
          description: |-
            * `idea` - Idea
            * `mvp` - MVP
            * `launch` - Launch
            * `scale` - Scale
            * `exit` - Exit
          x-spec-enum-id: 0a73ba3d72a52b97
          default: mvp
        fund_size:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - company_name
      - created_at
      - email
      - founded_year
      - fund_size
      - id
      - industry
      - location
      - team_size
      - updated_at
      - user
    InvestorRequest:
      type: object
      description: |-
        Serializer for the Investor model.
        Includes all fields defined in the abstract Company base class and Investor-specific fields.
      properties:
        industry:
          type: integer
        company_name:
          type: string
          minLength: 1
          maxLength: 254
        location:
          type: integer
        logo:
          type: string
          format: binary
          nullable: true
        description:
          type: string
        website:
          type: string
          format: uri
          maxLength: 200
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        founded_year:
          type: integer
          minimum: 1900
          maximum: 2025
        team_size:
          type: integer
          minimum: 1
        stage:
          enum:
          - idea
          - mvp
          - launch
          - scale
          - exit
          type: string
          description: |-
            * `idea` - Idea
            * `mvp` - MVP
            * `launch` - Launch
            * `scale` - Scale
            * `exit` - Exit
          x-spec-enum-id: 0a73ba3d72a52b97
          default: mvp
        fund_size:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
      required:
      - company_name
      - email
      - founded_year
      - fund_size
      - industry
      - location
      - team_size
...
    PatchedInvestorRequest:
      type: object
      description: |-
        Serializer for the Investor model.
        Includes all fields defined in the abstract Company base class and Investor-specific fields.
      properties:
        industry:
          type: integer
        company_name:
          type: string
          minLength: 1
          maxLength: 254
        location:
          type: integer
        logo:
          type: string
          format: binary
          nullable: true
        description:
          type: string
        website:
          type: string
          format: uri
          maxLength: 200
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
        founded_year:
          type: integer
          minimum: 1900
          maximum: 2025
        team_size:
          type: integer
          minimum: 1
        stage:
          enum:
          - idea
          - mvp
          - launch
          - scale
          - exit
          type: string
          description: |-
            * `idea` - Idea
            * `mvp` - MVP
            * `launch` - Launch
            * `scale` - Scale
            * `exit` - Exit
          x-spec-enum-id: 0a73ba3d72a52b97
          default: mvp
        fund_size:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
...